+++ packages/terra-button/tests/nightwatch/CompactButton.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import Button from '../../lib/Button';
+
+const CompactButton = () => <Button id="compactButton" text={'Compact'} />;
+
+export default CompactButton;