+++ packages/terra-alert/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Uplift to Button 2.0
+
+1.16.0 - (February 1, 2018)
+------------------
+### Changed
+* Minor version bump
 
 1.15.0 - (January 18, 2018)
 ------------------