+++ packages/terra-base/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+2.11.0 - (February 1, 2018)
+------------------
+### Added
+* Added terra-legacy-theme global styles to hyperlinks all over the components.
+
 2.10.0 - (January 18, 2018)
 ------------------
 ### Changed