+++ packages/terra-dynamic-grid/CHANGELOG.md
@@ -3,6 +3,10 @@ ChangeLog
 
 Unreleased
 -----------------
+
+1.12.0 - (February 1, 2018)
+------------------
+### Changed
 * Corrected example Region import path.
 
 1.11.0 - (January 18, 2018)