+++ packages/terra-i18n/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+1.14.0 - (February 1, 2018)
+------------------
+### Changed
+* Minor version bump
+
 1.13.0 - (January 18, 2018)
 ------------------
 ### Changed