+++ packages/terra-site/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Removed unknown `isDisabled` prop from InlineRadiosExample.jsx for terra-form-radio.
+* Update Collapsible Menu View and Toggle Button examples to use Button 2.0.
 ### Removed
 * terra-theme-provider examples
 * terra-app-delegate examples
@@ -14,6 +15,7 @@ Unreleased
 ------------------
 ### Changed
 * Add context to selectable list example
+* Update button group examples to use V2
 
 ### Added
 * Add terra-form-radio components