+++ packages/terra-status-view/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ------------------
+### Changed
+* Update to Button 2.0
+
+1.1.0 - (February 1, 2018)
+------------------
+### Changed
+* Minor version bump
 
 1.0.0 - (January 18, 2018)
 ------------------