+++ packages/terra-table/CHANGELOG.md
@@ -3,8 +3,11 @@ Changelog
 
 Unreleased
 ----------
+
+1.20.0 - (February 1, 2018)
+------------------
 ###Changed
-* Moved table sort indicator 
+* Moved table sort indicator
 
 1.19.0 - (January 18, 2018)
 ------------------